Liveness_Model/requirements.txt
opencv-python==4.4.0.40
tensorflow==2.5.0
pillow